chore: minor cleanups and refactors
