+++ app/src/main/java/com/example/trickcalculator/MainActivity.kt
@@ -38,11 +38,32 @@ class MainActivity : AppCompatActivity() {
         isDarkMode = isDarkMode()
     }
 
+    /**
+     * Pop most recent fragment from backstack.
+     */
     fun popBackStack() {
         val navHostFragment: NavHostFragment =
             supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
         val navController = navHostFragment.navController
         navController.popBackStack()
+    }
+
+    /**
+     * Run navigation action.
+     *
+     * @param actionResId [Int]: resource ID of action to run
+     * @param args [Bundle?]: arguments to pass with action
+     */
+    fun runNavAction(actionResId: Int, args: Bundle? = null) {
+        val navHostFragment =
+            supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
+        val navController = navHostFragment.navController
+
+        if (args == null) {
+            navController.navigate(actionResId)
+        } else {
+            navController.navigate(actionResId, args)
+        }
     }
 
     /**
