+++ app/src/main/java/com/example/trickcalculator/ui/BaseFragment.kt
@@ -1,6 +1,5 @@
 package com.example.trickcalculator.ui
 
-import android.os.Bundle
 import android.view.View
 import androidx.navigation.fragment.NavHostFragment
 import com.example.trickcalculator.MainActivity
@@ -22,6 +21,9 @@ abstract class BaseFragment : NavHostFragment() {
      */
     protected open var setActionBarOnClick: ((View) -> Unit)? = null
 
+    /**
+     * Re-add action bar settings when fragment is shown.
+     */
     override fun onResume() {
         super.onResume()
         runSetup()
@@ -48,22 +50,9 @@ abstract class BaseFragment : NavHostFragment() {
     }
 
     /**
-     * Run navigation action.
+     * Get current activity as [MainActivity].
      *
-     * @param actionResId [Int]: resource ID of action to run
-     * @param args [Bundle?]: arguments to pass with action
+     * @return [MainActivity]
      */
-    fun runNavAction(actionResId: Int, args: Bundle? = null) {
-        val navHostFragment =
-            requireActivity().supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
-        val navController = navHostFragment.navController
-
-        if (args == null) {
-            navController.navigate(actionResId)
-        } else {
-            navController.navigate(actionResId, args)
-        }
-    }
-
     fun requireMainActivity(): MainActivity = requireActivity() as MainActivity
 }
