+++ app/src/main/java/com/example/trickcalculator/ui/settings/components/settingsUtil.kt
@@ -39,32 +39,18 @@ fun getHistoryGroupValue(group: RadioGroup, buttons: List<RadioButton>): Int {
  * @param viewModel [SharedViewModel]: view model with settings fields
  * @param lifecycleOwner [LifecycleOwner]
  */
-private fun initObservers(
-    settingsUi: SettingsUI,
-    viewModel: SharedViewModel,
-    lifecycleOwner: LifecycleOwner
-) {
-    viewModel.applyDecimals.observe(lifecycleOwner) {
-        settingsUi.applyDecimalsSwitch.isChecked = it
-    }
+private fun initObservers(settingsUi: SettingsUI, viewModel: SharedViewModel, lifecycleOwner: LifecycleOwner) {
+    // update switches
+    viewModel.applyDecimals.observe(lifecycleOwner) { settingsUi.applyDecimalsSwitch.isChecked = it }
     viewModel.applyParens.observe(lifecycleOwner) { settingsUi.applyParensSwitch.isChecked = it }
     viewModel.clearOnError.observe(lifecycleOwner) { settingsUi.clearOnErrorSwitch.isChecked = it }
-    viewModel.showSettingsButton.observe(lifecycleOwner) {
-        settingsUi.settingsButtonSwitch.isChecked = it
-    }
-    viewModel.shuffleComputation.observe(lifecycleOwner) {
-        settingsUi.shuffleComputationSwitch.isChecked = it
-    }
-    viewModel.shuffleNumbers.observe(lifecycleOwner) {
-        settingsUi.shuffleNumbersSwitch.isChecked = it
-    }
-    viewModel.shuffleOperators.observe(lifecycleOwner) {
-        settingsUi.shuffleOperatorsSwitch.isChecked = it
-    }
+    viewModel.showSettingsButton.observe(lifecycleOwner) { settingsUi.settingsButtonSwitch.isChecked = it }
+    viewModel.shuffleComputation.observe(lifecycleOwner) { settingsUi.shuffleComputationSwitch.isChecked = it }
+    viewModel.shuffleNumbers.observe(lifecycleOwner) { settingsUi.shuffleNumbersSwitch.isChecked = it }
+    viewModel.shuffleOperators.observe(lifecycleOwner) { settingsUi.shuffleOperatorsSwitch.isChecked = it }
 
-    viewModel.historyRandomness.observe(lifecycleOwner) {
-        settingsUi.historyRadioButtons[it].isChecked = true
-    }
+    // update radio group
+    viewModel.historyRandomness.observe(lifecycleOwner) { settingsUi.historyRadioButtons[it].isChecked = true }
 }
 
 /**
