+++ app/src/main/res/layout/activity_main.xml
@@ -6,9 +6,7 @@
     android:layout_height="match_parent"
     tools:context=".MainActivity">
 
-    <include
-        android:id="@+id/actionBar"
-        layout="@layout/action_bar" />
+    <include android:id="@+id/actionBar" layout="@layout/action_bar" />
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/nav_host_fragment"
