+++ app/src/main/res/layout/component_close_button.xml
@@ -1,4 +1,6 @@
-<?xml version="1.0" encoding="utf-8"?><!-- Includes attributes for both ConstraintLayout and LinearLayout -->
+<?xml version="1.0" encoding="utf-8"?>
+
+<!-- Includes attributes for both ConstraintLayout and LinearLayout -->
 <ImageButton xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
