+++ app/src/main/res/layout/fragment_attributions.xml
@@ -6,9 +6,7 @@
     android:layout_height="match_parent"
     android:background="?backgroundColor">
 
-    <include
-        android:id="@+id/closeButton"
-        layout="@layout/component_close_button" />
+    <include android:id="@+id/closeButton" layout="@layout/component_close_button" />
 
     <!-- Flaticon info -->
     <TextView
