+++ app/src/main/res/layout/fragment_settings.xml
@@ -23,7 +23,7 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
 
-        <!-- Can't use close button component due to difference in constraints -->
+        <!-- Can't use close button component due to difference in layout constraints -->
         <ImageButton
             android:id="@+id/closeButton"
             android:layout_width="wrap_content"
