+++ app/src/main/res/values/attrs.xml
@@ -3,10 +3,6 @@
     <attr name="actionBarBackground" format="reference" />
     <attr name="actionBarTextColor" format="reference" />
 
-    <attr name="actionBarSwitchThumbColor" format="reference" />
-    <attr name="actionBarSwitchTrackCheckedColor" format="reference" />
-    <attr name="actionBarSwitchTrackUncheckedColor" format="reference" />
-
     <attr name="disabledForeground" format="reference" />
     <attr name="disabledBackground" format="reference" />
